requirements.txt
flask
flask-restful
flask-jwt-extended
Flask-SQLAlchemy
python-dotenv
psycopg2